+++ db/models/users.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import BigInteger, Column, Integer, String
 
 from db.models.base import Base
 
@@ -7,7 +7,7 @@ class Users(Base):
     """Model of user."""
     __tablename__ = 'users'
 
-    telegram_id = Column(Integer, primary_key=True)
+    telegram_id = Column(BigInteger, primary_key=True)
     name = Column(String)
     level = Column(Integer)
     main_language = Column(String)
